Part of WOR-510: decompose _parse_worker_telemetry CC 47→≤15 + split the telemetry-parse cluster out of watcher_helpers (S3776)
WOR-510 PR-b — the CC-47 monster, isolated per the ticket's sequencing.

1. CC reduction (behaviour-identical). _parse_worker_telemetry is a
   single-pass JSONL walk (WOR-466) whose complexity was a deeply-nested
   per-event branch. Decomposed into a _TelemetryAccum __slots__ struct
   (mirrors the WorkerTelemetry style; no new imports) + three verbatim
   per-event handlers (_handle_assistant/system/result_event). Same
   guards, same dispatch order, same accumulator mutations in the same
   sequence, same post-loop — values identical by construction.

2. File-size gate. The decomposition tipped watcher_helpers.py to 1212
   LOC, over the 1200 production BLOCK threshold (it was already ~1165 —
   at the edge). Rather than --no-verify (an honest edit that adds scope
   must not bypass the gate), the whole telemetry-parse cluster moved
   DOWN into watcher_log_parsing.py — its natural home: that module
   (layer 6, the WOR-403 parser split) already defines the
   _process_assistant_event / _resolve_usage_totals / _HOOK_VIOLATION_*
   the cluster calls, which watcher_helpers was importing back UP.

   Moved verbatim (439 LOC): WorkerBehavior, WorkerTelemetry,
   _TelemetryAccum, _handle_*_event, _parse_worker_telemetry,
   _parse_worker_behavior, _load_assistant_event, _walk_behavior_log,
   _update_input_tokens, _accumulate_content_block.

   watcher_helpers.py 1212→771 LOC (back under BLOCK, under the 800
   advisory too). watcher_log_parsing.py 373→815 (advisory tier only).
   The cluster is self-contained (verified: zero watcher_helpers-internal
   refs), so the move is layer-consistent — lint-imports 8 kept,
   0 broken, no .importlinter change. API stable: the 4 externally-
   imported symbols (WorkerBehavior/Telemetry, _parse_worker_telemetry/
   _behavior) are re-exported through watcher_helpers via the same facade
   pattern the WOR-403 split already uses → zero importer / test /
   mock-patch-string churn.

Telemetry feeds the metrics DB (this milestone's WOR-439 NULL-telemetry
saga shows the cost of a silent regression here): ruff / mypy /
lint-imports clean; 187 telemetry/behavior/parse_worker/finalize tests
pass; full unscoped pytest 2016 passed ×4 (identical count —
behaviour-identity).

Remaining: PR-c (S107 15-param dispatch-args dataclass + dispatch.py:52
CC) — own PR, high blast radius.

Part of WOR-510
