+++ app/core/watcher/watcher_helpers.py
@@ -9,22 +9,26 @@ from __future__ import annotations
 import json
 import logging
 import subprocess  # nosec B404
-from collections.abc import Iterable, Sequence
+from collections.abc import Sequence
 from pathlib import Path
 from typing import IO, Any
 
 from app.core.manifest import ExecutionManifest
 
+# WOR-510 PR-b: the telemetry-parse cluster (WorkerBehavior/Telemetry,
+# _parse_worker_telemetry/_behavior + their handlers) moved DOWN into
+# watcher_log_parsing to take watcher_helpers back under the 1200-LOC
+# block gate. Re-exported here for API stability — same facade pattern
+# the WOR-403 split already uses for the standalone parsers below.
 from .watcher_log_parsing import (
-    _HOOK_VIOLATION_TOKENS,
-    _is_violation_bash_command,
+    WorkerBehavior,
+    WorkerTelemetry,
     _parse_hook_trust_violations,
     _parse_worker_api_retries,
+    _parse_worker_behavior,
     _parse_worker_subagent_spawns,
+    _parse_worker_telemetry,
     _parse_worker_usage,
-    _process_assistant_event,
-    _process_compact_boundary,
-    _resolve_usage_totals,
     format_elapsed,
     format_token_count,
     format_worker_token_count,
@@ -53,7 +57,9 @@ __all__ = [
     "compute_vllm_metrics_delta",
     "picker_sort_key",
     "WorkerBehavior",
+    "WorkerTelemetry",
     "_parse_worker_behavior",
+    "_parse_worker_telemetry",
     "kv_concurrency_ceiling",
     "PRODUCTION_KV_CACHE_TOKENS",
     "COMPACTION_CONTEXT_CEILING",
@@ -769,397 +775,3 @@ def picker_sort_key(
             priority = 0
 
     return (parent_match, priority, candidate.get("id", ""))
-
-
-class WorkerBehavior:
-    """Per-worker behavior summary parsed from a stream-json log.
-
-    Concurrency-safe sibling to WOR-370 (server-side /metrics): every value
-    here is derived from the worker's own log file, so multiple concurrent
-    workers each produce their own attributable summary.
-
-    All fields are Optional to distinguish "definitely zero" from
-    "couldn't read the log" (None means the log was missing or unparseable
-    at the file-handle level; 0 / {} means the log was readable but the
-    feature didn't appear).
-    """
-
-    __slots__ = (
-        "turn_count",
-        "tool_calls_total",
-        "tool_calls_breakdown",
-        "thinking_blocks",
-        "thinking_chars_total",
-        "input_tokens_max",
-        "input_tokens_first",
-        "input_tokens_last",
-        "redundant_reads_count",
-    )
-
-    def __init__(
-        self,
-        turn_count: int | None,
-        tool_calls_total: int | None,
-        tool_calls_breakdown: dict[str, int] | None,
-        thinking_blocks: int | None,
-        thinking_chars_total: int | None,
-        input_tokens_max: int | None,
-        input_tokens_first: int | None,
-        input_tokens_last: int | None,
-        redundant_reads_count: int | None,
-    ) -> None:
-        self.turn_count = turn_count
-        self.tool_calls_total = tool_calls_total
-        self.tool_calls_breakdown = tool_calls_breakdown
-        self.thinking_blocks = thinking_blocks
-        self.thinking_chars_total = thinking_chars_total
-        self.input_tokens_max = input_tokens_max
-        self.input_tokens_first = input_tokens_first
-        self.input_tokens_last = input_tokens_last
-        self.redundant_reads_count = redundant_reads_count
-
-    @classmethod
-    def empty_unparseable(cls) -> "WorkerBehavior":
-        """Sentinel: log unreadable, so every field is None."""
-        return cls(None, None, None, None, None, None, None, None, None)
-
-    @classmethod
-    def empty_readable(cls) -> "WorkerBehavior":
-        """Sentinel: log was readable but had no assistant turns."""
-        return cls(0, 0, {}, 0, 0, None, None, None, 0)
-
-
-def _update_input_tokens(
-    inp: object,
-    first: int | None,
-    last: int | None,
-    cur_max: int | None,
-) -> tuple[int | None, int | None, int | None]:
-    """Fold a single usage.input_tokens value into the running (first, last, max)."""
-    if not isinstance(inp, int):
-        return first, last, cur_max
-    new_first = inp if first is None else first
-    new_last = inp
-    new_max = inp if cur_max is None else max(cur_max, inp)
-    return new_first, new_last, new_max
-
-
-def _accumulate_content_block(
-    block: dict[str, Any],
-    behavior_accum: dict[str, int],
-    tool_breakdown: dict[str, int],
-    read_counts: dict[str, int],
-) -> None:
-    """Update accumulators for one content block (thinking or tool_use)."""
-    btype = block.get("type")
-    if btype == "thinking":
-        behavior_accum["thinking_blocks"] += 1
-        text = block.get("thinking") or block.get("text") or ""
-        if isinstance(text, str):
-            behavior_accum["thinking_chars"] += len(text)
-    elif btype == "tool_use":
-        behavior_accum["tool_calls_total"] += 1
-        name = block.get("name") or "?"
-        tool_breakdown[name] = tool_breakdown.get(name, 0) + 1
-        if name == "Read":
-            fp = (block.get("input") or {}).get("file_path")
-            if isinstance(fp, str) and fp:
-                key = fp.replace("\\", "/")
-                read_counts[key] = read_counts.get(key, 0) + 1
-
-
-class WorkerTelemetry:
-    """Unified worker-log telemetry walked from one JSONL pass (WOR-466).
-
-    Replaces 5 separate `_parse_worker_*` walks at the finalize site with a
-    single traversal. The standalone parsers remain for other call sites
-    (heartbeat, ticket_status) that only need one slice.
-
-    Fields mirror the per-parser return types so the finalize_worker call
-    site can swap in cleanly:
-
-      * input_tokens / output_tokens / context_compactions / compact_duration_ms
-          -> matches `_parse_worker_usage`
-      * subagent_spawns          -> matches `_parse_worker_subagent_spawns`
-      * api_retries              -> matches `_parse_worker_api_retries`
-      * hook_trust_violations    -> matches `_parse_hook_trust_violations`
-      * behavior                 -> matches `_parse_worker_behavior`
-    """
-
-    __slots__ = (
-        "input_tokens",
-        "output_tokens",
-        "context_compactions",
-        "compact_duration_ms",
-        "subagent_spawns",
-        "api_retries",
-        "hook_trust_violations",
-        "behavior",
-    )
-
-    def __init__(
-        self,
-        input_tokens: int | None,
-        output_tokens: int | None,
-        context_compactions: int | None,
-        compact_duration_ms: int | None,
-        subagent_spawns: int | None,
-        api_retries: int | None,
-        hook_trust_violations: int | None,
-        behavior: WorkerBehavior,
-    ) -> None:
-        self.input_tokens = input_tokens
-        self.output_tokens = output_tokens
-        self.context_compactions = context_compactions
-        self.compact_duration_ms = compact_duration_ms
-        self.subagent_spawns = subagent_spawns
-        self.api_retries = api_retries
-        self.hook_trust_violations = hook_trust_violations
-        self.behavior = behavior
-
-    @classmethod
-    def empty_unparseable(cls) -> "WorkerTelemetry":
-        """Sentinel: log unreadable. Mirrors each parser's failure return."""
-        return cls(
-            input_tokens=None,
-            output_tokens=None,
-            context_compactions=None,
-            compact_duration_ms=None,
-            subagent_spawns=None,
-            api_retries=None,
-            hook_trust_violations=None,
-            behavior=WorkerBehavior.empty_unparseable(),
-        )
-
-
-def _parse_worker_telemetry(log_path: Path) -> WorkerTelemetry:
-    """Walk the worker JSONL log once and extract all telemetry (WOR-466).
-
-    On a typical 2-MB log this is ~5x faster than calling the 5 individual
-    parsers sequentially (one disk read + one json.loads per line vs five).
-    On large logs (~200 MB) the absolute saving is ~3-4 seconds.
-
-    Behavior matches the per-parser return types exactly — see WorkerTelemetry
-    docstring for the field mapping. On unreadable logs returns
-    WorkerTelemetry.empty_unparseable() (every field None / sentinel),
-    mirroring each old parser's failure return.
-    """
-    # Usage accumulators (mirror _parse_worker_usage's locals)
-    total_input = total_output = 0
-    has_assistant_usage = False
-    compact_count = compact_duration_ms = 0
-    last_input: int | None = None
-    last_output: int | None = None
-    content_chars = 0
-
-    # Counts
-    subagent_spawns = 0
-    api_retries = 0
-    hook_violations = 0
-
-    # Behavior accumulators (mirror _walk_behavior_log's locals)
-    turn_count = 0
-    behavior_accum = {
-        "thinking_blocks": 0,
-        "thinking_chars": 0,
-        "tool_calls_total": 0,
-    }
-    tool_breakdown: dict[str, int] = {}
-    input_tokens_first: int | None = None
-    input_tokens_last: int | None = None
-    input_tokens_max: int | None = None
-    read_counts: dict[str, int] = {}
-
-    try:
-        with log_path.open(encoding="utf-8") as f:
-            for raw in f:
-                line = raw.strip()
-                if not line:
-                    continue
-                try:
-                    obj = json.loads(line)
-                except json.JSONDecodeError:
-                    continue
-                if not isinstance(obj, dict):
-                    continue
-                obj_type = obj.get("type")
-
-                if obj_type == "assistant":
-                    # --- usage ---
-                    i_d, o_d, has_u, cc = _process_assistant_event(obj)
-                    total_input += i_d
-                    total_output += o_d
-                    has_assistant_usage = has_assistant_usage or has_u
-                    content_chars += cc
-
-                    # --- behavior ---
-                    turn_count += 1
-                    msg = obj.get("message") or {}
-                    usage = msg.get("usage") or {}
-                    (
-                        input_tokens_first,
-                        input_tokens_last,
-                        input_tokens_max,
-                    ) = _update_input_tokens(
-                        usage.get("input_tokens"),
-                        input_tokens_first,
-                        input_tokens_last,
-                        input_tokens_max,
-                    )
-
-                    # --- behavior content + spawns + hook violations ---
-                    for blk in msg.get("content") or []:
-                        if not isinstance(blk, dict):
-                            continue
-                        _accumulate_content_block(
-                            blk, behavior_accum, tool_breakdown, read_counts
-                        )
-                        if blk.get("type") == "tool_use":
-                            name = blk.get("name")
-                            if name == "Task":
-                                subagent_spawns += 1
-                            elif name == "Bash":
-                                token = _is_violation_bash_command(blk)
-                                if token and token in _HOOK_VIOLATION_TOKENS:
-                                    hook_violations += 1
-                elif obj_type == "system":
-                    subtype = obj.get("subtype")
-                    if subtype == "compact_boundary":
-                        c_d, d_d = _process_compact_boundary(obj)
-                        compact_count += c_d
-                        compact_duration_ms += d_d
-                    elif subtype == "api_retry":
-                        api_retries += 1
-                elif obj_type == "result":
-                    usage = obj.get("usage") or {}
-                    last_input = usage.get("input_tokens")
-                    last_output = usage.get("output_tokens")
-    except (OSError, ValueError) as exc:
-        logger.warning(
-            "Failed to read %s for unified telemetry — falling back to "
-            "empty_unparseable; %s",
-            log_path,
-            exc,
-            exc_info=True,
-        )
-        return WorkerTelemetry.empty_unparseable()
-
-    input_tok, output_tok, compactions, compact_dur = _resolve_usage_totals(
-        total_input,
-        total_output,
-        has_assistant_usage,
-        compact_count,
-        compact_duration_ms,
-        last_input,
-        last_output,
-        content_chars,
-    )
-
-    if turn_count == 0:
-        behavior = WorkerBehavior.empty_readable()
-    else:
-        redundant = sum(1 for n in read_counts.values() if n > 2)
-        behavior = WorkerBehavior(
-            turn_count=turn_count,
-            tool_calls_total=behavior_accum["tool_calls_total"],
-            tool_calls_breakdown=tool_breakdown,
-            thinking_blocks=behavior_accum["thinking_blocks"],
-            thinking_chars_total=behavior_accum["thinking_chars"],
-            input_tokens_max=input_tokens_max,
-            input_tokens_first=input_tokens_first,
-            input_tokens_last=input_tokens_last,
-            redundant_reads_count=redundant,
-        )
-
-    return WorkerTelemetry(
-        input_tokens=input_tok,
-        output_tokens=output_tok,
-        context_compactions=compactions,
-        compact_duration_ms=compact_dur,
-        subagent_spawns=subagent_spawns,
-        api_retries=api_retries,
-        hook_trust_violations=hook_violations,
-        behavior=behavior,
-    )
-
-
-def _parse_worker_behavior(log_path: Path) -> WorkerBehavior:
-    """Extract per-session behavior signals from the stream-json worker log.
-
-    Counts turns, tool calls + breakdown, thinking blocks + chars, and the
-    input_tokens trajectory (first/last/max). Derives redundant_reads_count
-    from per-file Read counts. Returns ``empty_unparseable()`` on read/parse
-    failure; ``empty_readable()`` if parsed but no assistant turns.
-    """
-    try:
-        with log_path.open(encoding="utf-8") as f:
-            return _walk_behavior_log(f)
-    except (OSError, ValueError) as exc:
-        logger.warning(
-            "Failed to read %s — behaviour telemetry columns will be NULL; %s",
-            log_path,
-            exc,
-            exc_info=True,
-        )
-        return WorkerBehavior.empty_unparseable()
-
-
-def _load_assistant_event(line: str) -> dict[str, Any] | None:
-    """Parse one JSONL line; return dict only if it is an assistant event."""
-    stripped = line.strip()
-    if not stripped:
-        return None
-    try:
-        obj = json.loads(stripped)
-    except json.JSONDecodeError:
-        return None
-    if not isinstance(obj, dict) or obj.get("type") != "assistant":
-        return None
-    return obj
-
-
-def _walk_behavior_log(lines: Iterable[str]) -> WorkerBehavior:
-    """Aggregate per-session behavior signals from a stream of JSONL lines."""
-    turn_count = 0
-    behavior_accum = {
-        "thinking_blocks": 0,
-        "thinking_chars": 0,
-        "tool_calls_total": 0,
-    }
-    tool_breakdown: dict[str, int] = {}
-    input_tokens_first: int | None = None
-    input_tokens_last: int | None = None
-    input_tokens_max: int | None = None
-    read_counts: dict[str, int] = {}
-    for raw in lines:
-        obj = _load_assistant_event(raw)
-        if obj is None:
-            continue
-        turn_count += 1
-        msg = obj.get("message") or {}
-        usage = msg.get("usage") or {}
-        input_tokens_first, input_tokens_last, input_tokens_max = _update_input_tokens(
-            usage.get("input_tokens"),
-            input_tokens_first,
-            input_tokens_last,
-            input_tokens_max,
-        )
-        for blk in msg.get("content") or []:
-            if isinstance(blk, dict):
-                _accumulate_content_block(
-                    blk, behavior_accum, tool_breakdown, read_counts
-                )
-    if turn_count == 0:
-        return WorkerBehavior.empty_readable()
-    redundant = sum(1 for n in read_counts.values() if n > 2)
-    return WorkerBehavior(
-        turn_count=turn_count,
-        tool_calls_total=behavior_accum["tool_calls_total"],
-        tool_calls_breakdown=tool_breakdown,
-        thinking_blocks=behavior_accum["thinking_blocks"],
-        thinking_chars_total=behavior_accum["thinking_chars"],
-        input_tokens_max=input_tokens_max,
-        input_tokens_first=input_tokens_first,
-        input_tokens_last=input_tokens_last,
-        redundant_reads_count=redundant,
-    )
